pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "hedge-fund-13f-tracker"
version = "0.1.0"
description = "Monitor institutional investor holdings via SEC 13F filings"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1",
    "rich>=13.0",
    "requests>=2.31",
    "beautifulsoup4>=4.12",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "ruff>=0.1",
]
web = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
]

[project.scripts]
tracker = "src.cli.main:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.ruff]
line-length = 100
target-version = "py311"
